Refactor files/folders layout and naming convention for easier understanding/
