+++ client/src/App.js
@@ -7,10 +7,9 @@ import UserProfileManager from "./components/user/UserAuthentication";
 import Profile from "./components/user/Profile";
 import PrivateRoute from "./components/routing/privateRoute";
 import SolveQuestion from "./components/userCollaboration/SolveQuestion.js";
-import QuestionDetails from "./components/userCollaboration/QuestionDetails";
 import "./App.css";
 
-/* To enable sending of cookies on each request globally*/
+/* To enable sending of cookies on each request globally */
 axios.defaults.withCredentials = true;
 
 const App = () => (
@@ -21,7 +20,6 @@ const App = () => (
       <Route path="/login" element={<UserProfileManager />} />
       <Route path="/profile" element={<PrivateRoute component={Profile}/>} />
       <Route path="/solve-question" element={<PrivateRoute component={SolveQuestion}/>} />
-      <Route path="/question-testing" element={<QuestionDetails/>}/>
     </Routes>
   </Router>
 );
