+++ client/src/components/userCollaboration/collabViewComponents/infoBar.css
