+++ client/src/components/userCollaboration/solveQuestion.css
@@ -1,3 +1,4 @@
 .solve-question-page {
   height: 100vh;
+  max-width: 100vw;
 }
